requirements.dev.txt
# Common
-r requirements.txt
# Stubs
django-stubs[compatible-mypy]
types-python-dateutil